# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
paths:
...
  /api/phone-verification:
    get:
      summary: 'Send verification code'
      operationId: sendVerificationCode
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Verification code has been sent, please check your messenger.'
                properties:
                  message:
                    type: string
                    example: 'Verification code has been sent, please check your messenger.'
        400:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Verification code has been sent, please check your messenger.'
                properties:
                  message:
                    type: string
                    example: 'Verification code has been sent, please check your messenger.'
      tags:
        - Authentication
  /api/phone-verification-resend:
    post:
      summary: 'Resend verification code'
      operationId: resendVerificationCode
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Verification code sent.'
                properties:
                  message:
                    type: string
                    example: 'Verification code sent.'
        400:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Verification code sent.'
                properties:
                  message:
                    type: string
                    example: 'Verification code sent.'
      tags:
        - Authentication
  /api/phone-verify:
    post:
      summary: 'Verify code'
      operationId: verifyCode
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Verification successful!'
                properties:
                  message:
                    type: string
                    example: 'Verification successful!'
        400:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Invalid verification code. Please try again.'
                properties:
                  message:
                    type: string
                    example: 'Invalid verification code. Please try again.'
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                verification_code:
                  type: integer
                  description: 'The 4-digit verification code.'
                  example: 1234
              required:
                - verification_code
...